+++ CHANGELOG.md
@@ -24,7 +24,7 @@
 
 - change project structure to use it as **git submodule**
 - rename functionality to **branch protection** from *annotational marker check*
-- rename functionality to **ensure file modification** from *ensure file change*
+- rename functionality to **ensure file modified** from *ensure file change*
 
 ### Deprecated
 ### Removed